localON/pyproject.toml
[project]
name = "localon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiomysql>=0.2.0",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.1",
    "sqlalchemy>=2.0.36",
    "uvicorn[standard]>=0.30.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.2",
    "alembic>=1.18.4",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
